move discoverTopology to topo package
