+++ pkg/kubelet/cpumanager/topo/topo.go
@@ -16,6 +16,12 @@ limitations under the License.
 
 package topo
 
+import (
+	"fmt"
+
+	cadvisorapi "github.com/google/cadvisor/info/v1"
+)
+
 //CPU  - logical CPU, cadvisor - thread
 //Core - physical CPU, cadvisor - Core
 //Socket - socket, cadvisor - Node
@@ -30,4 +36,40 @@ type CPUTopology struct {
 type CPUInfo struct {
 	SocketId int
 	CoreId   int
+}
+
+func DiscoverTopology(machineInfo *cadvisorapi.MachineInfo) (*CPUTopology, error) {
+
+	if machineInfo.NumCores == 0 {
+		return nil, fmt.Errorf("could not detect number of cpus")
+	}
+
+	CPUtopoDetails := make(map[int]CPUInfo)
+
+	numCPUs := machineInfo.NumCores
+	htEnabled := false
+	numPhysicalCores := 0
+	for _, socket := range machineInfo.Topology {
+		numPhysicalCores += len(socket.Cores)
+		for _, core := range socket.Cores {
+			for _, cpu := range core.Threads {
+				CPUtopoDetails[cpu] = CPUInfo{
+					CoreId:   core.Id,
+					SocketId: socket.Id,
+				}
+				// a little bit naive
+				if !htEnabled && len(core.Threads) != 1 {
+					htEnabled = true
+				}
+			}
+		}
+	}
+
+	return &CPUTopology{
+		NumCPUs:        numCPUs,
+		NumSockets:     len(machineInfo.Topology),
+		NumCores:       numPhysicalCores,
+		HyperThreading: htEnabled,
+		CPUtopoDetails: CPUtopoDetails,
+	}, nil
 }
